Implement GET /api/events/:id route
Server retrieves an event by its ID number from the database. Server returns data to client.

Server gets the ID number from the GEt route's params.
