+++ package.json
@@ -7,8 +7,11 @@
   "engines": {
     "node": ">=6.13.0"
   },
-  "scripts": {},
+  "scripts": {
+    "start": "nodemon server/app.js"
+  },
   "dependencies": {
+    "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "sqlite3": "^4.0.6"
